requirements.txt
# Model building requirements
numpy>=1.26,<2.0
pandas>=2.0,<3.0
joblib==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
pytest==7.4.2
setuptools == 67.8.0
wheel==0.38.4
fastapi==0.103.0
pydantic==2.4.2
uvicorn==0.23.2
gunicorn==21.2.0
Pillow==11.3.0



# Monitoring
prometheus-fastapi-instrumentator==6.1.0